pyproject.toml
[project]
name = "klas-research-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "isort>=5.13.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.3",
    "sqlfluff>=3.2.5",
]
